test(provenance,oracle): add tests for issues #37, #38, #39, #40
Issue #38: move provenance tests to test.rs, add test_double_initialization
         and test_mint_without_initialization
Issue #39: add test_mint_certificate (single mint + field assertions),
         test_mint_multiple_certificates, test_get_nonexistent_certificate
Issue #40: add test_prevent_duplicate_manifest_hash
Issue #37: verify_tee_hash tests (success, invalid hash, registry failure)
         already present in oracle/src/lib.rs
