+++ contracts/provenance/src/lib.rs
@@ -112,90 +112,4 @@ impl ProvenanceContract {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use soroban_sdk::{testutils::Address as _, Env, String};
-
-    fn s(env: &Env, v: &str) -> String {
-        String::from_str(env, v)
-    }
-
-    #[test]
-    fn test_double_initialization() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let cid = env.register_contract(None, ProvenanceContract);
-        let client = ProvenanceContractClient::new(&env, &cid);
-        let oracle = Address::generate(&env);
-        client.initialize(&oracle);
-        assert!(client.try_initialize(&oracle).is_err());
-    }
-
-    #[test]
-    fn test_mint_without_initialization() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let cid = env.register_contract(None, ProvenanceContract);
-        let client = ProvenanceContractClient::new(&env, &cid);
-        let to = Address::generate(&env);
-        assert!(client
-            .try_mint(&s(&env, "sid"), &s(&env, "mhash"), &s(&env, "ahash"), &to)
-            .is_err());
-    }
-
-    #[test]
-    fn test_mint_multiple_certificates() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let cid = env.register_contract(None, ProvenanceContract);
-        let client = ProvenanceContractClient::new(&env, &cid);
-        let oracle = Address::generate(&env);
-        let to = Address::generate(&env);
-        client.initialize(&oracle);
-        assert_eq!(client.mint(&s(&env, "sid"), &s(&env, "mh1"), &s(&env, "ah"), &to), 1);
-        assert_eq!(client.mint(&s(&env, "sid"), &s(&env, "mh2"), &s(&env, "ah"), &to), 2);
-        assert_eq!(client.mint(&s(&env, "sid"), &s(&env, "mh3"), &s(&env, "ah"), &to), 3);
-    }
-
-    // #13
-    #[test]
-    fn test_prevent_duplicate_manifest_hash() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let cid = env.register_contract(None, ProvenanceContract);
-        let client = ProvenanceContractClient::new(&env, &cid);
-        let oracle = Address::generate(&env);
-        let to = Address::generate(&env);
-        client.initialize(&oracle);
-        client.mint(&s(&env, "sid"), &s(&env, "mhash"), &s(&env, "ahash"), &to);
-        assert!(client
-            .try_mint(&s(&env, "sid"), &s(&env, "mhash"), &s(&env, "ahash"), &to)
-            .is_err());
-    }
-
-    // #15
-    #[test]
-    fn test_certificate_minted_event() {
-        use soroban_sdk::testutils::Events as _;
-        let env = Env::default();
-        env.mock_all_auths();
-        let cid = env.register_contract(None, ProvenanceContract);
-        let client = ProvenanceContractClient::new(&env, &cid);
-        let oracle = Address::generate(&env);
-        let to = Address::generate(&env);
-        client.initialize(&oracle);
-        client.mint(&s(&env, "sid"), &s(&env, "mhash"), &s(&env, "ahash"), &to);
-        let events = env.events().all();
-        assert!(!events.is_empty());
-    }
-
-    // #16
-    #[test]
-    fn test_get_nonexistent_certificate() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let cid = env.register_contract(None, ProvenanceContract);
-        let client = ProvenanceContractClient::new(&env, &cid);
-        assert!(client.try_get_certificate(&999u64).is_err());
-    }
-}
+mod test;
